// packages/openapi-ts/src/types/schema.ts
export interface Extensable {
	[key: `x-${string}`]: any;
}
// ...
export interface ServerObject extends Extensable {
	/** REQUIRED. A URL to the target host. This URL supports Server Variables and MAY be relative, to indicate that the host location is relative to the location where the OpenAPI document is being served. Variable substitutions will be made when a variable is named in {brackets}. */
	url: string;
	/** An optional string describing the host designated by the URL. CommonMark syntax MAY be used for rich text representation. */
	description: string;
	/** A map between a variable name and its value. The value is used for substitution in the server’s URL template. */
	variables: { [name: string]: ServerVariableObject };
}
// ...
export interface PathsObject {
	[pathname: string]: PathItemObject | ReferenceObject; // note: paths object does support $refs; the schema just defines it in a weird way
}
// ...
export interface PathItemObject extends Extensable {
	/** A definition of a GET operation on this path. */
	get?: OperationObject | ReferenceObject;
	/** A definition of a PUT operation on this path. */
	put?: OperationObject | ReferenceObject;
	/** A definition of a POST operation on this path. */
	post?: OperationObject | ReferenceObject;
	/** A definition of a DELETE operation on this path. */
	delete?: OperationObject | ReferenceObject;
	/** A definition of a OPTIONS operation on this path. */
	options?: OperationObject | ReferenceObject;
	/** A definition of a HEAD operation on this path. */
	head?: OperationObject | ReferenceObject;
	/** A definition of a PATCH operation on this path. */
	patch?: OperationObject | ReferenceObject;
	/** A definition of a TRACE operation on this path. */
	trace?: OperationObject | ReferenceObject;
	/** An alternative server array to service all operations in this path. */
	servers?: ServerObject[];
	/** A list of parameters that are applicable for all the operations described under this path. These parameters can be overridden at the operation level, but cannot be removed there. The list MUST NOT include duplicated parameters. A unique parameter is defined by a combination of a name and location. The list can use the Reference Object to link to parameters that are defined at the OpenAPI Object’s components/parameters. */
	parameters?: (ParameterObject | ReferenceObject)[];
}

/**
 * [4.8.10] Operation Object
 * Describes a single API operation on a path.
 */
export interface OperationObject extends Extensable {
	/** A list of tags for API documentation control. Tags can be used for logical grouping of operations by resources or any other qualifier. */
	tags?: string[];
	/** A short summary of what the operation does. */
	summary?: string;
	/** A verbose explanation of the operation behavior. CommonMark syntax MAY be used for rich text representation. */
	description?: string;
	/** Additional external documentation for this operation. */
	externalDocs?: ExternalDocumentationObject;
	/** Unique string used to identify the operation. The id MUST be unique among all operations described in the API. The operationId value is case-sensitive. Tools and libraries MAY use the operationId to uniquely identify an operation, therefore, it is RECOMMENDED to follow common programming naming conventions. */
	operationId?: string;
	/** A list of parameters that are applicable for this operation. If a parameter is already defined at the Path Item, the new definition will override it but can never remove it. The list MUST NOT include duplicated parameters. A unique parameter is defined by a combination of a name and location. The list can use the Reference Object to link to parameters that are defined at the OpenAPI Object’s components/parameters. */
	parameters?: (ParameterObject | ReferenceObject)[];
	/** The request body applicable for this operation. The requestBody is fully supported in HTTP methods where the HTTP 1.1 specification [RFC7231] has explicitly defined semantics for request bodies. In other cases where the HTTP spec is vague (such as GET, HEAD and DELETE), requestBody is permitted but does not have well-defined semantics and SHOULD be avoided if possible. */
	requestBody?: RequestBodyObject | ReferenceObject;
	/** The list of possible responses as they are returned from executing this operation. */
	responses?: ResponsesObject;
	/** A map of possible out-of band callbacks related to the parent operation. The key is a unique identifier for the Callback Object. Each value in the map is a Callback Object that describes a request that may be initiated by the API provider and the expected responses. */
	callbacks?: Record<string, CallbackObject | ReferenceObject>;
	/** Declares this operation to be deprecated. Consumers SHOULD refrain from usage of the declared operation. Default value is false. */
	deprecated?: boolean;
	/** A declaration of which security mechanisms can be used for this operation. The list of values includes alternative security requirement objects that can be used. Only one of the security requirement objects need to be satisfied to authorize a request. To make security optional, an empty security requirement ({}) can be included in the array. This definition overrides any declared top-level security. To remove a top-level security declaration, an empty array can be used. */
	security?: SecurityRequirementObject[];
	/** An alternative server array to service this operation. If an alternative server object is specified at the Path Item Object or Root level, it will be overridden by this value. */
	servers?: ServerObject[];
}
// ...
export interface ParameterObject extends Extensable {
	/**
	 * REQUIRED. The name of the parameter. Parameter names are case sensitive.
	 *
	 * - If `in` is `"path"`, the `name` field MUST correspond to a template expression occurring within the path field in the Paths Object. See Path Templating for further information.
	 * - If `in` is `"header"` and the `name` field is `"Accept"`, `"Content-Type"` or `"Authorization"`, the parameter definition SHALL be ignored.
	 * - For all other cases, the `name` corresponds to the parameter name used by the `in` property.
	 */
	name: string;
	/** REQUIRED. The location of the parameter. Possible values are "query", "header", "path" or "cookie".*/
	in: "query" | "header" | "path" | "cookie";
	/** A brief description of the parameter. This could contain examples of use. CommonMark syntax MAY be used for rich text representation. */
	description?: string;
	/** Determines whether this parameter is mandatory. If the parameter location is "path", this property is REQUIRED and its value MUST be true. Otherwise, the property MAY be included and its default value is false. */
	required?: boolean;
	/** Specifies that a parameter is deprecated and SHOULD be transitioned out of usage. Default value is false. */
	deprecated?: boolean;
	/** Sets the ability to pass empty-valued parameters. This is valid only for query parameters and allows sending a parameter with an empty value. Default value is false. If style is used, and if behavior is n/a (cannot be serialized), the value of allowEmptyValue SHALL be ignored. Use of this property is NOT RECOMMENDED, as it is likely to be removed in a later revision. */
	allowEmptyValue?: boolean;
	/** Describes how the parameter value will be serialized depending on the type of the parameter value. Default values (based on value of in): for query - form; for path - simple; for header - simple; for cookie - form. */
	style?: string;
	/** When this is true, parameter values of type `array` or `object` generate separate parameters for each value of the array or key-value pair of the map. For other types of parameters this property has no effect. When `style` is `form`, the default value is `true`. For all other styles, the default value is `false`. */
	explode?: boolean;
	/** Determines whether the parameter value SHOULD allow reserved characters, as defined by [RFC3986] `:/?#[]@!$&'()*+,;=` to be included without percent-encoding. This property only applies to parameters with an `in` value of `query`. The default value is `false`. */
	allowReserved?: boolean;
	/** The schema defining the type used for the parameter. */
	schema?: SchemaObject;
	/** Example of the parameter’s potential value. */
	example?: any;
	/** Examples of the parameter’s potential value. */
	examples?: { [name: string]: ExampleObject | ReferenceObject };
	/** A map containing the representations for the parameter. */
	content?: { [contentType: string]: MediaTypeObject | ReferenceObject };
}
// ...
export interface ReferenceObject extends Extensable {
	/** REQUIRED. The reference identifier. This MUST be in the form of a URI. */
	$ref: string;
	/** A short summary which by default SHOULD override that of the referenced component. If the referenced object-type does not allow a summary field, then this field has no effect. */
	summary?: string;
	/** A description which by default SHOULD override that of the referenced component. CommonMark syntax MAY be used for rich text representation. If the referenced object-type does not allow a description field, then this field has no effect. */
	description?: string;
}